+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [UNRELEASED]
+### Fixed/Changed
+ - Improved matching of undefined locale. Ensures language tag extensions are passed through the locale matcher and resource bundle builder.
+
 ## [0.14.3]
 ### Fixed/Changed
  - Bug: if locale matcher found an exact match, extensions would not be applied to the resulting tag.